made bounce mechanic symmetric need to set bounce length to get back to board piece

+++ src/enumcs.cpp
@@ -2,15 +2,15 @@
 
 INTERACTION_FLAG operator|(INTERACTION_FLAG lhs, INTERACTION_FLAG rhs) {
     return static_cast<INTERACTION_FLAG>(
-        static_cast<__uint128_t>(lhs) |
-        static_cast<__uint128_t>(rhs)
+        static_cast<int>(lhs) |
+        static_cast<int>(rhs)
     );
 }
 
 INTERACTION_FLAG operator&(INTERACTION_FLAG lhs, INTERACTION_FLAG rhs) {
     return static_cast<INTERACTION_FLAG>(
-        static_cast<__uint128_t>(lhs) &
-        static_cast<__uint128_t>(rhs)
+        static_cast<int>(lhs) &
+        static_cast<int>(rhs)
     );
 }
